requirements.txt
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.9
langgraph==0.2.45
langchain-chroma==0.1.4
chromadb==0.5.18
sqlalchemy==2.0.35
gradio==4.44.1
fastmcp==0.4.1
pypdf==5.1.0
python-dotenv==1.0.1
openai==1.54.4
tiktoken==0.8.0
pandas==2.2.3
faker==33.0.0
uvicorn==0.32.1
pydantic-settings==2.6.1
pydantic==2.9.2
tabulate==0.9.0